Compatibility update for aws deployment
Made some changes to file structure and reading methods to work inside the final compiled JAR.  Reworked the fetch requests on the front end to work from the same root directory as the server.  Reverted the java environment version and spring boot version for compatibility with elastic beanstalk java 11 environment.